+++ src/storage/src/storage/bidi.rs
@@ -30,15 +30,8 @@ use crate::request_options::RequestOptions;
 use gaxi::grpc::tonic::{Extensions, Response as TonicResponse, Result as TonicResult};
 use tokio::sync::mpsc::Receiver;
 
-#[cfg(google_cloud_unstable_grpc_rust)]
-pub(crate) type GrpcClient = gaxi::grpc::GrpcRustClient;
-#[cfg(google_cloud_unstable_grpc_rust)]
-pub(crate) type GrpcStream = gaxi::grpc::GrpcRustStreaming<BidiReadObjectResponse>;
-
-#[cfg(not(google_cloud_unstable_grpc_rust))]
 pub(crate) type GrpcClient = gaxi::grpc::Client;
-#[cfg(not(google_cloud_unstable_grpc_rust))]
-pub(crate) type GrpcStream = gaxi::grpc::tonic::Streaming<BidiReadObjectResponse>;
+pub(crate) type GrpcStream = gaxi::grpc::Streaming<BidiReadObjectResponse>;
 
 pub use super::open_object::OpenObject;
