+++ src/storage/src/storage/bidi_write.rs
@@ -34,15 +34,8 @@ use gaxi::grpc::tonic::{Extensions, Response as TonicResponse, Result as TonicRe
 use std::future::Future;
 use tokio::sync::mpsc::Receiver;
 
-#[cfg(google_cloud_unstable_grpc_rust)]
-pub(crate) type GrpcClient = gaxi::grpc::GrpcRustClient;
-#[cfg(google_cloud_unstable_grpc_rust)]
-pub(crate) type GrpcStream = gaxi::grpc::GrpcRustStreaming<BidiWriteObjectResponse>;
-
-#[cfg(not(google_cloud_unstable_grpc_rust))]
 pub(crate) type GrpcClient = gaxi::grpc::Client;
-#[cfg(not(google_cloud_unstable_grpc_rust))]
-pub(crate) type GrpcStream = gaxi::grpc::tonic::Streaming<BidiWriteObjectResponse>;
+pub(crate) type GrpcStream = gaxi::grpc::Streaming<BidiWriteObjectResponse>;
 
 /// A trait to mock `Streaming<T>` in the unit tests.
 ///
